Working all fine with scriptable object

+++ PrefabManipulation/Assets/Scripts/AIStats.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "AI/Stats")]
 public class AIStats : ScriptableObject
 {
-    public float hp;
+    public int hp;
     public float speed;
 
     public void Save()
